+++ ES.Data/Models/PartnerTypeModel.cs
@@ -6,7 +6,7 @@
         {
             
         }
-        public long Id { get; set; }
+        public short Id { get; set; }
         public string Description { get; set; }
     }
 }
